fix all platform errors
